# RestWithASP/Collections/REST_With_ASP.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST_With_ASP",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Book"
    },
    {
      "name": "Person"
    }
  ],
  "paths": {
    "/api/auth/v{version}/signin": {
      "post": {
        "summary": "/api/auth/v{version}/signin",
        "x-apidog-folder": "Auth",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344558-run",
        "security": []
      }
    },
    "/api/book/v{version}": {
      "get": {
        "summary": "/api/book/v{version}",
        "x-apidog-folder": "Book",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookVO"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344559-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "/api/book/v{version}",
        "x-apidog-folder": "Book",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookVO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344560-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "/api/book/v{version}",
        "x-apidog-folder": "Book",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookVO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344561-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/book/v{version}/{id}": {
      "get": {
        "summary": "/api/book/v{version}/{id}",
        "x-apidog-folder": "Book",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookVO"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344562-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "/api/book/v{version}/{id}",
        "x-apidog-folder": "Book",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344563-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/person/v{version}": {
      "get": {
        "summary": "/api/person/v{version}",
        "x-apidog-folder": "Person",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonVO"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344564-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "/api/person/v{version}",
        "x-apidog-folder": "Person",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonVO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344565-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "/api/person/v{version}",
        "x-apidog-folder": "Person",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonVO"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonVO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344566-run",
        "security": []
      }
    },
    "/api/person/v{version}/{id}": {
      "get": {
        "summary": "/api/person/v{version}/{id}",
        "x-apidog-folder": "Person",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonVO"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344567-run",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "/api/person/v{version}/{id}",
        "x-apidog-folder": "Person",
        "x-apidog-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "Person"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-orders": [],
                  "x-apidog-ignore-properties": []
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/932981/apis/api-17344568-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "PersonVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HyperMediaLink"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "x-apidog-orders": [
          "id",
          "firstName",
          "lastName",
          "address",
          "gender",
          "links"
        ],
        "x-apidog-ignore-properties": [],
        "x-apidog-folder": "Schemas"
      },
      "HyperMediaLink": {
        "type": "object",
        "properties": {
          "rel": {
            "type": "string",
            "nullable": true
          },
          "href": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "action": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "x-apidog-orders": [
          "rel",
          "href",
          "type",
          "action"
        ],
        "x-apidog-ignore-properties": [],
        "x-apidog-folder": "Schemas"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {
          "type": "string"
        },
        "x-apidog-orders": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "x-apidog-ignore-properties": [],
        "x-apidog-folder": "Schemas"
      },
      "BookVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "launch_date": {
            "type": "string",
            "format": "date-time"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HyperMediaLink"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "x-apidog-orders": [
          "id",
          "title",
          "author",
          "price",
          "launch_date",
          "links"
        ],
        "x-apidog-ignore-properties": [],
        "x-apidog-folder": "Schemas"
      },
      "UserVO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "x-apidog-orders": [
          "password",
          "userName"
        ],
        "x-apidog-ignore-properties": [],
        "x-apidog-folder": "Schemas"
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://localhost:44370",
      "description": "REST_With_ASP"
    }
  ]
}